Delete club_id mocks
