+++ mocks/handlers/api/club/[clubID]/get.ts
@@ -1,48 +1,14 @@
-import { rest } from 'msw'
 import serverApi from '@/mocks/handlers/serverApi'
+import { rest } from 'msw'
+import { data } from './data'
 
-const getClubDetail = rest.get(serverApi('/club/:clubID'), (_req, res, ctx) => {
-  return res(
-    ctx.json({
-      id: 1,
-      type: 'MAJOR',
-      bannerImg: 'https://bit.ly/3EsznxZ',
-      name: '동아리 이름',
-      content: 'tjfaud',
-      contact: '윤지빈#5193',
-      teacher: '선생님이름', // optional
-      isOpened: true,
-      notionLink: 'https://bit.ly/3EsznxZ',
-      activityImgs: [
-        'https://bit.ly/3EsznxZ',
-        'https://bit.ly/3EsznxZ',
-        'https://bit.ly/3EsznxZ',
-        'https://bit.ly/3EsznxZ',
-      ],
-      head: {
-        uuid: 'uuuid',
-        email: 's21073',
-        name: 'name',
-        grade: 1, // is
-        classNum: 2,
-        num: 14,
-        userImg: 'url', // optional
-      },
-      member: [
-        {
-          uuid: 'uuid',
-          email: 's21074',
-          name: 'name',
-          grade: 1,
-          classNum: 2,
-          num: 14,
-          userImg: 'url', // optional
-        },
-      ],
-      scope: 'HEAD',
-      isApplied: true,
-    })
-  )
-})
+const getClubDetail = rest.get(
+  serverApi('/club/:clubID'),
+  async (req, res, ctx) => {
+    const clubID = req.params.clubID.toString()
+    if (!clubID) return res(ctx.status(400))
+    return res(ctx.json(data.find((i) => i.id.toString().includes(clubID))))
+  }
+)
 
 export default getClubDetail
